+++ components/keyedit/api/src/main/java/com/flipperdevices/keyedit/api/KeyEditApi.kt
@@ -1,8 +1,0 @@
-package com.flipperdevices.keyedit.api
-
-import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
-import com.github.terrakok.cicerone.Screen
-
-interface KeyEditApi {
-    fun getScreen(flipperKeyPath: FlipperKeyPath, title: String? = null): Screen
-}