+++ components/widget/api/src/main/java/com/flipperdevices/widget/api/WidgetFeatureEntry.kt
@@ -8,6 +8,5 @@ interface WidgetFeatureEntry : ComposableFeatureEntry {
     override val ROUTE: FeatureScreenRootRoute
         get() = FeatureScreenRootRoute.WIDGET
 
-    fun getWidgetScreen(widgetId: Int): String
     fun getWidgetScreenByDeeplink(deeplink: Deeplink): String
 }