+++ components/widget/api/src/main/java/com/flipperdevices/widget/api/WidgetScreenApi.kt
@@ -1,7 +1,0 @@
-package com.flipperdevices.widget.api
-
-import com.github.terrakok.cicerone.Screen
-
-interface WidgetScreenApi {
-    fun getWidgetOptionsScreen(widgetId: Int): Screen
-}